Remove supertest packages: It forces us to do many tests with promises when that's not necessary.
For example, once we have made a request and the request data is returned, the test for status code should not require a promise. This is a problem because if we forget to return the promise in a cucumber step, the test silently passes because the promise is never executed. I prefer tests that immediately fail or pass.