Don't use dates in resources, always use strings. Reasoning:
Ok so I thought about this some more and I think in general we should not put non-serializable types in resources. For example, we would not put a field like { users: Set<UserResource> }  because sets are not part of the JSON standard; we would always use an array here. When building the resource, TypeScript will force us to transform our set to an array.
The same should be the case for dates. Since Date is not part of the JSON standard, declaring a Date in a resource would be incorrect, because the actually JSON returned will be a string. So we should always declare our dates as strings in resources, because this is how they will be transmitted anyways. This also means we can/must ensure ourselves that the dates we transmit are always ISO-compliant.